UWPAzureIoTHubSDK
Azure IoTHub UWP SDK facilitates the service endpoint to send and
receive cloud messages to the device end points. This SDK can be used
for developing universal windows platform (UWP) in c# for windows 10 OS
, developed by e-con systems and contributed to Azure IoT. This library
uses AMQPNetlite to communicate with the Azure IoT hub. A sample
application is provided to know the usage of the APIs.